added support for migrations with flyway, custom json serialization, content negotiation and hateoas implementation
