--- 10_VersioningAPIs/src/main/java/br/com/ervandil/vo/v1/PersonVO.java
@@ -1,73 +1,0 @@
-package br.com.ervandil.vo.v1;
-
-import java.io.Serial;
-import java.io.Serializable;
-import java.util.Objects;
-
-public class PersonVO implements Serializable {
-
-    @Serial
-    private static final long serialVersionUID = 1L;
-
-    private Long id;
-    private String firstName;
-    private String lastName;
-    private String address;
-    private String gender;
-
-    public PersonVO() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        PersonVO person = (PersonVO) o;
-        return Objects.equals(id, person.id) && Objects.equals(firstName, person.firstName) && Objects.equals(lastName, person.lastName) && Objects.equals(address, person.address) && Objects.equals(gender, person.gender);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, firstName, lastName, address, gender);
-    }
-}
