--- 10_VersioningAPIs/src/main/resources/application.yml
@@ -1,14 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/rest_with_spring_boot_erudio?useTimezone=true&serverTimezone=UTC
-    username: root
-    password: admin123
-  jpa:
-    open-in-view: false
-    hibernate:
-      ddl-auto: update
-#    properties:
-#      hibernate:
-#        dialect: org.hibernate.dialect.MySQLDialect
-    show-sql: false
