+++ 11_MigrationSupportWithFlyway/src/main/java/br/com/ervandil/controllers/PersonController.java
@@ -1,0 +1,56 @@
+package br.com.ervandil.controllers;
+
+
+import br.com.ervandil.services.PersonServices;
+import br.com.ervandil.util.MediaType;
+import br.com.ervandil.vo.v1.PersonVO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/person/v1")
+public class PersonController {
+
+    @Autowired
+    private PersonServices service;
+
+    @GetMapping(value = "/{id}", produces = {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.APPLICATION_YML})
+    public PersonVO findById(@PathVariable(value = "id") Long id) {
+        return service.findById(id);
+    }
+
+    @GetMapping(produces = {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.APPLICATION_YML})
+    public List<PersonVO> findAll() {
+        return service.findAll();
+    }
+
+    @PostMapping(
+            produces = {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.APPLICATION_YML},
+            consumes = {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.APPLICATION_YML})
+    public PersonVO create(@RequestBody PersonVO person) {
+        return service.create(person);
+    }
+
+//    @PostMapping(value = "/v2",
+//            produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE},
+//            consumes = MediaType.APPLICATION_JSON_VALUE)
+//    public PersonVOV2 createV2(@RequestBody PersonVOV2 person) {
+//        return service.createV2(person);
+//    }
+
+    @PutMapping(
+            produces = {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.APPLICATION_YML},
+            consumes = {MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, MediaType.APPLICATION_YML})
+    public PersonVO update(@RequestBody PersonVO person) {
+        return service.update(person);
+    }
+
+    @DeleteMapping(value = "/{id}")
+    public ResponseEntity<?> delete(@PathVariable(value = "id") Long id) {
+        service.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
