+++ 11_MigrationSupportWithFlyway/src/main/java/br/com/ervandil/model/Person.java
