+++ 11_MigrationSupportWithFlyway/src/main/java/br/com/ervandil/serialization/converter/YamlJackson2HttpMesageConverter.java
@@ -1,0 +1,14 @@
+package br.com.ervandil.serialization.converter;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
+import org.springframework.http.MediaType;
+import org.springframework.http.converter.json.AbstractJackson2HttpMessageConverter;
+
+public class YamlJackson2HttpMesageConverter extends AbstractJackson2HttpMessageConverter {
+    public YamlJackson2HttpMesageConverter() {
+        super(new YAMLMapper().setSerializationInclusion(JsonInclude.Include.NON_NULL), MediaType.parseMediaType("application/x-yaml"));
+    }
+
+
+}
