+++ 11_MigrationSupportWithFlyway/src/main/java/br/com/ervandil/util/MediaType.java
@@ -1,0 +1,8 @@
+package br.com.ervandil.util;
+
+public class MediaType {
+
+    public static final String APPLICATION_YML = "application/x-yaml";
+    public static final String APPLICATION_JSON = "application/json";
+    public static final String APPLICATION_XML = "application/xml";
+}
