+++ 11_MigrationSupportWithFlyway/src/main/resources/db/migration/V1__Create_Table_person.sql
@@ -1,0 +1,9 @@
+CREATE TABLE `person` (
+  `id` bigint NOT NULL AUTO_INCREMENT,
+  `first_name` varchar(80) NOT NULL,
+  `last_name` varchar(80) NOT NULL,
+  `address` varchar(100) NOT NULL,
+  `gender` varchar(6) NOT NULL,
+  `birth_day` datetime(6) DEFAULT NULL,
+  PRIMARY KEY (`id`)
+)
